+++ build.zig
@@ -424,6 +424,7 @@ pub fn build(b: *std.Build) void {
         "tests/cli/local_import_test.zig",
         "tests/cli/lsp_test.zig",
         "tests/cli/sqlite_test.zig",
+        "tests/cli/orm_test.zig",
         "tests/cli/shared_mem_test.zig",
         "tests/cli/install_test.zig",
         "tests/cli/help_screen_test.zig",